accounts/requirements.txt
Django==4.2.5
django-widget-tweaks==1.5.0
Pillow==10.1.0